requirements.txt
langchain
langchainhub
langgraph
langchain_core
langchain_community
langchain_tavily
langchain_groq
black
isort
python-dotenv
faiss-cpu 
sentence-transformers 
pypdf
unstructured[pdf]
langchain_text_splitters
langchain_huggingface
streamlit
crewai
crewai_tools
qdrant_client
LiteLLM